Backed out 2 changesets (bug 1739448) for causing geckoview failures on 1246775-1.html. CLOSED TREE
Backed out changeset 7b82837e2f56 (bug 1739448)
Backed out changeset feac99f1dae3 (bug 1739448)

UltraBlame original commit: 829b201706904131278d53c67f9091e34a2c56a7

+++ gfx/skia/skia/include/core/SkPaint.h
@@ -268,7 +268,7 @@ public:
     float getAlphaf() const { return fColor4f.fA; }
 
     
-    uint8_t getAlpha() const { return uint8_t(sk_float_round2int(this->getAlphaf() * 255)); }
+    uint8_t getAlpha() const { return sk_float_round2int(this->getAlphaf() * 255); }
